+++ package.json
@@ -20,8 +20,8 @@
     "postinstall": "npm run typings && npm run build",
     "start": "cross-env NODE_ENV=production node server/node-server.js",
     "pretest": "npm run lint",
-    "test": "ts-babel-node ./node_modules/.bin/_mocha --opts ./test-utils/mocha.opts",
-    "test:watch": "ts-babel-node ./node_modules/.bin/_mocha --opts ./test-utils/mocha.opts --watch --watch-extensions ts,tsx",
+    "test": "./node_modules/.bin/_mocha --opts ./test/utils/mocha.opts",
+    "test:watch": "./node_modules/.bin/_mocha --opts ./test/utils/mocha.opts --watch --watch-extensions ts,tsx",
     "typings": "rimraf typings/ && typings install"
   },
   "keywords": [
