--- src/actions/counter.ts
@@ -1,14 +1,0 @@
-import { INCREMENT_COUNTER, DECREMENT_COUNTER } from '../constants';
-
-
-export function increment() {
-  return {
-    type: INCREMENT_COUNTER,
-  };
-}
-
-export function decrement() {
-  return {
-    type: DECREMENT_COUNTER,
-  };
-}
