+++ src/containers/counter-page.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 const connect = require('react-redux').connect;
 
-import { increment, decrement } from '../actions/counter';
+import { increment, decrement } from '../feature/counter';
 import Counter from '../components/counter';
 import Container from '../components/container';
 
@@ -27,7 +27,7 @@ function mapDispatchToProps(dispatch) {
 class CounterPage extends React.Component<ICounterPageProps, void> {
   render() {
     const { counter, increaseCounter, decreaseCounter } = this.props;
- 
+
     return <Container size={2} center>
       <h2 className="center caps" id="qa-counter-heading">Counter</h2>
