+++ test/utils/mocha.opts
@@ -1,0 +1,4 @@
+--require ts-node/register
+--reporter dot
+test/**/*.ts
+
